Model Commands - Collision
I hate the way both of these are implemented but I was going insane trying to fix them so conceded that an inferior implementation will do for now

+++ ResourceChangedEvent.h
@@ -7,8 +7,8 @@ namespace hk
 	struct ResourceChangedEvent
 	{
 		std::string id;
-		float		previous_amount = 0.0f;
-		float		new_amount = 0.0f;
-		float		change_in_amount = 0.0f;
+		double		previous_amount = 0.0f;
+		double		new_amount = 0.0f;
+		double		change_in_amount = 0.0f;
 	};
 }
